+++ test.js
@@ -1,0 +1,5 @@
+const {viewDepartments, viewRoles, viewEmployees, addDepartment, addRole, addEmployee, updateEmployee} = require('./query');
+
+viewDepartments();
+viewEmployees();
+viewRoles();
